requirements.txt
# requirements.txt
# Python 3.10+ recommended

torch
torchvision
numpy
matplotli